# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API provided by DTI project",
        "title": "Domain Threat Intelligence API",
        "contact": {
            "name": "Yaroslav Lysak",
            "url": "https://t.me/Qvineox"
        },
        "version": "0.0.4"
    },
    "paths": {
...
        "/jobs/job": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes single job",
                "tags": [
                    "Jobs"
                ],
                "summary": "Delete single job by UUID",
                "parameters": [
                    {
                        "description": "record UUID to delete",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routing.byUUIDParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/success.DatabaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
...
        "/jobs/jobs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns list of jobs by filter",
                "tags": [
                    "Jobs"
                ],
                "summary": "Get jobs by filter",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "multi",
                        "description": "Job type IDs",
                        "name": "types[]",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Job status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Job priority",
                        "name": "priority",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Created by user with ID",
                        "name": "created_by",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Only finished jobs",
                        "name": "is_finished",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created timestamp is after",
                        "name": "created_after",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created timestamp is before",
                        "name": "created_before",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Query limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Query offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jobEntities.Job"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
...
        "/scanning/queue/job": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Accepts and adds new scanning job to queue",
                "tags": [
                    "Queue"
                ],
                "summary": "Enqueue scanning job",
                "parameters": [
                    {
                        "description": "New job to queue",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/jobEntities.JobCreateParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/routing.queuedJob"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes single job from queue",
                "tags": [
                    "Queue"
                ],
                "summary": "Delete single job from queue by UUID",
                "parameters": [
                    {
                        "description": "job UUID to delete",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routing.removeFromQueueByUUIDParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/scanning/queue/jobs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns list of jobs from queue",
                "tags": [
                    "Queue"
                ],
                "summary": "Enqueued scanning jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jobEntities.Job"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "error.APIError": {
            "type": "object",
            "properties": {
                "ErrorCode": {
                    "$ref": "#/definitions/error.CodeError"
                },
                "ErrorMessage": {
                    "type": "string"
                },
                "ErrorModule": {
                    "type": "string"
                },
                "StatusCode": {
                    "type": "integer"
                }
            }
        },
        "error.CodeError": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15
            ],
            "x-enum-varnames": [
                "EncodingErrorCode",
                "DecodingErrorCode",
                "IncorrectParamsErrorCode",
                "InsufficientParamsErrorCode",
                "DatabaseErrorCode",
                "DatabaseEntityNotFoundCode",
                "FileDecodingErrorCode",
                "FileReadingErrorCode",
                "FileExtensionNotSupportedErrorCode",
                "NotImplementedErrorCode",
                "FileProcessingErrorCode",
                "InternalUnidentifiedErrorCode",
                "AuthFailedErrorCode",
                "AuthPermissionInsufficientErrorCode",
                "JobQueueErrorCode"
            ]
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "jobEntities.DNSDirectives": {
            "type": "object",
            "properties": {
                "Timings": {
                    "$ref": "#/definitions/jobEntities.DirectiveTimings"
                }
            }
        },
        "jobEntities.DirectiveTimings": {
            "type": "object",
            "required": [
                "Delay",
                "Reties",
                "Timeout"
            ],
            "properties": {
                "Delay": {
                    "type": "integer"
                },
                "Reties": {
                    "type": "integer"
                },
                "Timeout": {
                    "type": "integer"
                }
            }
        },
        "jobEntities.Directives": {
            "type": "object",
            "properties": {
                "DNSDirectives": {
                    "$ref": "#/definitions/jobEntities.DNSDirectives"
                },
                "DiscoveryDirectives": {
                    "$ref": "#/definitions/jobEntities.DiscoveryDirectives"
                },
                "NMAPDirectives": {
                    "$ref": "#/definitions/jobEntities.NMAPDirectives"
                },
                "OpenSourceScanDirectives": {
                    "$ref": "#/definitions/jobEntities.OSSDirectives"
                },
                "SpiderDirectives": {
                    "$ref": "#/definitions/jobEntities.SpiderDirectives"
                },
                "WhoISDirectives": {
                    "$ref": "#/definitions/jobEntities.WhoISDirectives"
                }
            }
        },
        "jobEntities.DiscoveryDirectives": {
            "type": "object",
            "properties": {
                "Ports": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "Silent": {
                    "type": "boolean"
                },
                "Timings": {
                    "$ref": "#/definitions/jobEntities.DirectiveTimings"
                }
            }
        },
        "jobEntities.Job": {
            "type": "object",
            "properties": {
                "DequeuedTimes": {
                    "type": "integer"
                },
                "Directives": {
                    "$ref": "#/definitions/jobEntities.Directives"
                },
                "Meta": {
                    "$ref": "#/definitions/jobEntities.Metadata"
                },
                "NodeScans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/networkEntities.NetworkNodeScan"
                    }
                },
                "Payload": {
                    "$ref": "#/definitions/jobEntities.Payload"
                }
            }
        },
        "jobEntities.JobCreateParams": {
            "type": "object",
            "required": [
                "Targets"
            ],
            "properties": {
                "CreatedByUserID": {
                    "type": "integer"
                },
                "Delay": {
                    "type": "integer"
                },
                "Exceptions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "Priority": {
                    "enum": [
                        0,
                        1,
                        2,
                        3
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/jobEntities.JobPriority"
                        }
                    ]
                },
                "Private": {
                    "type": "boolean"
                },
                "Providers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/jobEntities.SupportedOSSProvider"
                    }
                },
                "Retries": {
                    "type": "integer"
                },
                "Targets": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "Timout": {
                    "type": "integer"
                },
                "Type": {
                    "enum": [
                        0,
                        1,
                        2,
                        3,
                        4,
                        5
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/jobEntities.JobType"
                        }
                    ]
                },
                "UseHomeBound": {
                    "type": "boolean"
                },
                "Weight": {
                    "type": "integer"
                }
            }
        },
        "jobEntities.JobPriority": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-comments": {
                "JOB_PRIORITY_CRITICAL": "job must be executed instantly",
                "JOB_PRIORITY_HIGH": "job must be executed after current (stack mode)",
                "JOB_PRIORITY_LOW": "job should be executed lastly in order (queue mode)",
                "JOB_PRIORITY_MEDIUM": "job should be executed with higher priority"
            },
            "x-enum-varnames": [
                "JOB_PRIORITY_CRITICAL",
                "JOB_PRIORITY_HIGH",
                "JOB_PRIORITY_MEDIUM",
                "JOB_PRIORITY_LOW"
            ]
        },
        "jobEntities.JobStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7
            ],
            "x-enum-comments": {
                "JOB_STATUS_CANCELLED": "job was cancelled by user",
                "JOB_STATUS_DONE": "job finished execution and saved",
                "JOB_STATUS_ERROR": "job stopped with error from API or scanners (can be multiple errors, with threshold)",
                "JOB_STATUS_FINISHING": "clearing and sending data",
                "JOB_STATUS_PANIC": "internal exception",
                "JOB_STATUS_PENDING": "not yet started",
                "JOB_STATUS_STARTING": "calculating tasks, creating required structures",
                "JOB_STATUS_WORKING": "executing tasks"
            },
            "x-enum-varnames": [
                "JOB_STATUS_PENDING",
                "JOB_STATUS_STARTING",
                "JOB_STATUS_WORKING",
                "JOB_STATUS_FINISHING",
                "JOB_STATUS_DONE",
                "JOB_STATUS_ERROR",
                "JOB_STATUS_PANIC",
                "JOB_STATUS_CANCELLED"
            ]
        },
        "jobEntities.JobType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5
            ],
            "x-enum-varnames": [
                "JOB_TYPE_OSS",
                "JOB_TYPE_NMAP",
                "JOB_TYPE_WHOIS",
                "JOB_TYPE_DNS",
                "JOB_TYPE_DISCOVERY",
                "JOB_TYPE_SPIDER"
            ]
        },
        "jobEntities.Metadata": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "CreatedBy": {
                    "$ref": "#/definitions/userEntities.PlatformUser"
                },
                "CreatedByID": {
                    "type": "integer"
                },
                "DeletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "Error": {
                    "type": "string"
                },
                "FinishedAt": {
                    "type": "string"
                },
                "Priority": {
                    "$ref": "#/definitions/jobEntities.JobPriority"
                },
                "StartedAt": {
                    "type": "string"
                },
                "Status": {
                    "$ref": "#/definitions/jobEntities.JobStatus"
                },
                "TasksLeft": {
                    "type": "integer"
                },
                "Type": {
                    "$ref": "#/definitions/jobEntities.JobType"
                },
                "UUID": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                },
                "Weight": {
                    "type": "integer"
                }
            }
        },
        "jobEntities.NMAPDirectives": {
            "type": "object",
            "properties": {
                "Timings": {
                    "$ref": "#/definitions/jobEntities.DirectiveTimings"
                }
            }
        },
        "jobEntities.OSSDirectives": {
            "type": "object",
            "properties": {
                "Providers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/jobEntities.SupportedOSSProvider"
                    }
                },
                "Timings": {
                    "$ref": "#/definitions/jobEntities.DirectiveTimings"
                }
            }
        },
        "jobEntities.Payload": {
            "type": "object",
            "properties": {
                "Exceptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/jobEntities.Target"
                    }
                },
                "Targets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/jobEntities.Target"
                    }
                }
            }
        },
        "jobEntities.SpiderDirectives": {
            "type": "object",
            "properties": {
                "Depth": {
                    "type": "integer"
                },
                "Timings": {
                    "$ref": "#/definitions/jobEntities.DirectiveTimings"
                }
            }
        },
        "jobEntities.SupportedOSSProvider": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "OSS_PROVIDER_VIRUS_TOTAL",
                "OSS_PROVIDER_IP_QUALITY_SCORE",
                "OSS_PROVIDER_CROWD_SEC",
                "OSS_PROVIDER_SHODAN",
                "OSS_PROVIDER_IP_WHO_IS"
            ]
        },
        "jobEntities.Target": {
            "type": "object",
            "properties": {
                "Host": {
                    "type": "string"
                },
                "Type": {
                    "$ref": "#/definitions/jobEntities.TargetType"
                }
            }
        },
        "jobEntities.TargetType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "HOST_TYPE_CIDR",
                "HOST_TYPE_DOMAIN",
                "HOST_TYPE_URL",
                "HOST_TYPE_EMAIL"
            ]
        },
        "jobEntities.WhoISDirectives": {
            "type": "object",
            "properties": {
                "Timings": {
                    "$ref": "#/definitions/jobEntities.DirectiveTimings"
                }
            }
        },
...
        "networkEntities.NetworkNode": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "DiscoveredAt": {
                    "description": "Network node discovery timestamp, when was this node first found",
                    "allOf": [
                        {
                            "$ref": "#/definitions/sql.NullTime"
                        }
                    ]
                },
                "Identity": {
                    "description": "Network node unique identity, can be any address or URI. Must be unique.",
                    "type": "string"
                },
                "NodeType": {
                    "$ref": "#/definitions/networkEntities.NetworkNodeType"
                },
                "NodeTypeId": {
                    "type": "integer"
                },
                "Scans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/networkEntities.NetworkNodeScan"
                    }
                },
                "UUID": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "networkEntities.NetworkNodeScan": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "Data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "DeletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "ID": {
                    "type": "integer"
                },
                "IsComplete": {
                    "type": "boolean"
                },
                "JobUUID": {
                    "description": "Defines in which job scan result was created",
                    "type": "string"
                },
                "Node": {
                    "description": "Defines parent node, scan object belongs to node object (many-to-one)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/networkEntities.NetworkNode"
                        }
                    ]
                },
                "NodeUUID": {
                    "type": "string"
                },
                "Type": {
                    "$ref": "#/definitions/networkEntities.NetworkNodeScanType"
                },
                "TypeID": {
                    "type": "integer"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "networkEntities.NetworkNodeScanType": {
            "type": "object",
            "properties": {
                "DeletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "Description": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                }
            }
        },
        "networkEntities.NetworkNodeType": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "Description": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
...
        "routing.byUUIDParams": {
            "type": "object",
            "required": [
                "UUID"
            ],
            "properties": {
                "UUID": {
                    "type": "string"
                }
            }
        },
...
        "routing.queuedJob": {
            "type": "object",
            "properties": {
                "UUID": {
                    "type": "string"
                }
            }
        },
        "routing.removeFromQueueByUUIDParams": {
            "type": "object",
            "required": [
                "UUID"
            ],
            "properties": {
                "Force": {
                    "type": "boolean"
                },
                "UUID": {
                    "type": "string"
                }
            }
        },
...
        "sql.NullTime": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "success.DatabaseResponse": {
            "type": "object",
            "properties": {
                "RowsAffected": {
                    "type": "integer"
                },
                "StatusCode": {
                    "type": "integer"
                },
                "Warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "userEntities.PlatformUser": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "Email": {
                    "type": "string"
                },
                "FullName": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "IsActive": {
                    "type": "boolean"
                },
                "Login": {
                    "type": "string"
                },
                "Permissions": {
                    "description": "Defines which roles user has",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userEntities.PlatformUserPermission"
                    }
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "userEntities.PlatformUserPermission": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "Description": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "IsActive": {
                    "type": "boolean"
                },
                "Name": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-api-Key",
            "in": "header"
        }
    }
}